@@ -1,15 +1,16 @@
-﻿using Microsoft.AspNetCore.Mvc;
+﻿using Microsoft.AspNetCore.Authorization;
+using Microsoft.AspNetCore.Mvc;
 
-namespace WebCaster.API.Controllers
+namespace WebCaster.API.Controllers;
+
+[Route("api/system")]
+[Tags("System")]
+[ApiController]
+public class System : ControllerBase
 {
-    [Route("api/system")]
-    [ApiController]
-    public class System : ControllerBase
+    [HttpGet("ping"), AllowAnonymous]
+    public ActionResult<DateTime> Ping()
     {
-        [HttpGet("ping")]
-        public ActionResult<DateTime> Ping()
-        {
-            return Ok(DateTime.Now.ToUniversalTime());
-        }
+        return Ok(DateTime.Now.ToUniversalTime());
     }
 }
